Fixed scheduled parsing and file init.py
