+++ parser_main.py
@@ -1,9 +1,10 @@
 import sqlalchemy as sa
 import sqlalchemy.orm as so
-from webapp import app, db
-from webapp.models import User, Query, Parser, ResultParser
+from webapp import create_app
+from webapp.db import db
+from webapp.models import User, Parser, ResultParser
 
 
-@app.shell_context_processor
+@create_app.shell_context_processor
 def make_shell_context():
-  return {'sa': sa, 'so': so, 'db': db, 'User': User, 'Query': Query, 'Parser': Parser, 'ResultParser': ResultParser}
+  return {'sa': sa, 'so': so, 'db': db, 'User': User, 'Parser': Parser, 'ResultParser': ResultParser}
