+++ webapp/models.py
@@ -2,10 +2,10 @@ from datetime import datetime, timezone
 from typing import Optional
 import sqlalchemy as sa
 import sqlalchemy.orm as so
-from webapp import db
+from webapp.db import db
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import UserMixin
-from webapp import login
+# from webapp import login
 
 
 class User(UserMixin, db.Model):
@@ -13,7 +13,7 @@ class User(UserMixin, db.Model):
     username: so.Mapped[str] = so.mapped_column(sa.String(64), index=True, unique=True)
     email: so.Mapped[str] = so.mapped_column(sa.String(120), index=True, unique=True)
     password_hash: so.Mapped[Optional[str]] = so.mapped_column(sa.String(256))
-    querys: so.WriteOnlyMapped['Query'] = so.relationship(back_populates='author') 
+    # querys: so.WriteOnlyMapped['Query'] = so.relationship(back_populates='author') 
 
     def set_password(self, password):
         self.password_hash = generate_password_hash(password)
@@ -21,28 +21,28 @@ class User(UserMixin, db.Model):
     def check_password(self, password):
         return check_password_hash(self.password_hash, password)
 
-    @property
-    def is_admin(self):
-        return self.role == 'admin'
+    # @property
+    # def is_admin(self):
+    #     return self.role == 'admin'
 
     def __repr__(self):
         return '<User name={} id={}>'.format(self.username, self.id)
     
 
-class Query(db.Model):
-    id: so.Mapped[int] = so.mapped_column(primary_key=True)
-    body: so.Mapped[str] = so.mapped_column(sa.String(256))
-    timestamp: so.Mapped[datetime] = so.mapped_column(index=True, default=lambda: datetime.now(timezone.utc))
-    user_id: so.Mapped[int] = so.mapped_column(sa.ForeignKey(User.id), index=True)
-    author: so.Mapped[User] = so.relationship(back_populates='querys')
+# class Query(db.Model):
+#     id: so.Mapped[int] = so.mapped_column(primary_key=True)
+#     body: so.Mapped[str] = so.mapped_column(sa.String(256))
+#     timestamp: so.Mapped[datetime] = so.mapped_column(index=True, default=lambda: datetime.now(timezone.utc))
+#     user_id: so.Mapped[int] = so.mapped_column(sa.ForeignKey(User.id), index=True)
+#     author: so.Mapped[User] = so.relationship(back_populates='querys')
 
-    def __repr__(self):
-        return '<Query {}>'.format(self.body)
+#     def __repr__(self):
+#         return '<Query {}>'.format(self.body)
     
 
-    @login.user_loader
-    def load_user(id):
-        return db.session.get(User, int(id))
+    # @login.user_loader
+    # def load_user(id):
+    #     return db.session.get(User, int(id))
     
 
 class Parser(db.Model):
